feat: add funtionality to update users list after adding user, change files structure
